dashboard/requirements.txt
flask
dash
dash_bootstrap_components
plotly
configparser
mysql-connector
